+++ test/test.carbone.js
@@ -11,6 +11,7 @@ var spawn = require('child_process').spawn;
 var execSync = require('child_process').execSync;
 var pdfjsLib = require('pdfjs-dist/legacy/build/pdf.js');
 const helperTest = require('./helper');
+const pdf2json  = require('pdf2json');
 
 describe('Carbone', function () {
 
@@ -3434,6 +3435,46 @@ describe('Carbone', function () {
         var _result = execSync('lsof -p '+process.pid).toString();
         assert.equal(/test_word_render_A\.docx/.test(_result), false);
         done();
+      });
+    });
+    it('should accept visio files', function (done) {
+      const _data = {
+        id      : 'field_1',
+        test    : 'hello world',
+        message : 'the earth'
+      };
+      carbone.render('test_visio_simple.vsdx', _data, { renderPrefix : 'prefix', convertTo : 'vsdx' }, function (err, resultFilePath) {
+        assert.equal(err, null);
+        assert.strictEqual(path.dirname(resultFilePath), params.renderPath);
+        fs.mkdirSync(testPath, parseInt('0755',8));
+        var _unzipPath = path.join(testPath, 'unzipVisio1');
+        unzipSystem(resultFilePath, _unzipPath, function (err, files) {
+          var _page1 = files['visio/pages/page1.xml'];
+          var _page2 = files['visio/pages/page2.xml'];
+          assert.notEqual(_page1.indexOf('field_1'), -1);
+          assert.notEqual(_page1.indexOf('the earth'), -1);
+          assert.notEqual(_page2.indexOf('hello world'), -1);
+          fs.unlinkSync(resultFilePath);
+          done();
+        });
+      });
+    });
+    it('should accept to convert visio files to PDF', function (done) {
+      const _data = {
+        id      : 'field_1',
+        test    : 'hello world',
+        message : 'the earth'
+      };
+      carbone.render('test_visio_simple.vsdx', _data, { renderPrefix : 'prefix', convertTo : 'pdf' }, function (err, resultFilePath) {
+        helper.assert(err+'', 'null');
+        const _pdf = new pdf2json();
+        _pdf.loadPDF(resultFilePath);
+        _pdf.on('pdfParser_dataReady', pdfData => {
+          helper.assert(pdfData.Meta.PDFFormatVersion !== undefined, true);
+          helper.assert(pdfData.Pages.length, 2);
+          helper.assert(pdfData.Pages[0].Texts[0].R[0].T, 'the%20earth%20');
+          done();
+        });
       });
     });
     it('should be fast to render a document without conversion', function (done) {